fix media control and organize code
